% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_requests.R
\name{client-requests}
\alias{client-requests}
\alias{new_session}
\title{High level client requests.}
\usage{
new_session(client, clone = NULL)
}
\arguments{
\item{client}{Client function created by a call to \code{\link{client}}
constructor.}

\item{clone}{Id of the session to clone.}
}
\description{
These are high level wrapers around \code{\link{sync_request}} to request
the operations provided by default nREPL middleware.
}

