% Generated by roxygen2 (4.0.2): do not edit by hand
\name{client}
\alias{client}
\alias{combine_responses}
\alias{connect}
\alias{sync_request}
\alias{sync_request0}
\title{Low level functions for client connections.}
\usage{
connect(port = 4005, host = "localhost", transport_fn = transport_bencode)

client(transport)

sync_request(client, op, ...)

sync_request0(client, op, ...)

combine_responses(response_list)
}
\arguments{
\item{port}{Port number of a running nREPL instance .}

\item{host}{Host address. Defaults to "localhost".}

\item{transport_fn}{Constructor that returns a transport connection
object. See \code{\link{transport}}.}

\item{transport}{Transport connection object returned by \code{connect} or a
transport constructor such as \code{\link{transport_bencode}}.}

\item{client}{Client function produced by \code{client} constructor.}

\item{op}{Name of the requested operation (for example, "eval", "describe",
"load-file" etc)}

\item{...}{Key-value pairs of the request to be sent to nREPL server.}

\item{response_list}{List of \code{bendict} objects to be combined together.}
}
\description{
Functions to perform the connect -> request -> receive -> close cycle.
}
\details{
In order to communicate to nREPL server you need first to create a
connection object with \code{connect}, then feed it to \code{client}
constructor which returns a function of variable arity.

Receive responses from nREPL one at a time by calling the client function
with no arguments. Send messages to nREPL server by calling client function
with key-value named arguments representing a message.

\code{sync_request} sends a requests and waits till the all the messages
associates with the current request have been received. Then returns a list
of the \code{bendict} responses. \code{sync_request0} is like
\code{sync_request} but combines all the responses from \code{sync_request}
into one \code{bendict} object.

As R has no multi-thread support, asynchronous requests are not yet
implemented.
}
\seealso{
\code{\link{transport}}
}

