% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc_middleware.R
\name{epc_middleware}
\alias{epc_middleware}
\alias{error_for.EPC}
\alias{mw_epc_describe}
\alias{mw_epc_echo}
\alias{mw_epc_session}
\alias{resp_for.EPC}
\title{EPC middleware}
\usage{
\method{resp_for}{EPC}(msg, ..., op = q(return), .lst = list())

\method{error_for}{EPC}(msg, message = "", ...)

mw_epc_describe(handler)

mw_epc_session(handler)

mw_epc_echo(handler)
}
\arguments{
\item{handler}{Handler function. See \code{\link{middleware}}.}
}
\description{
EPC middleware
}

