% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/middleware.R
\name{q}
\alias{backToTop}
\alias{backToTopError}
\alias{error_for}
\alias{q}
\alias{resp_for}
\alias{test_middleware}
\alias{unknown_op}
\title{Various utils}
\usage{
q(x)

error_for(msg, message = "", ..., additional_status = list())

resp_for(msg, ..., .lst = list())

unknown_op(tr, msg)

backToTop()

backToTopError(message, ..., status = list())

test_middleware(mw, ...)
}
\arguments{
\item{msg}{Error message to send back to client on backToTop condition}

\item{message}{error message to be sent through}

\item{mw}{Middleware for testing}
}
\description{
Various utils
}

