% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mw_describe}
\alias{middlewares}
\alias{mw_describe}
\alias{mw_eval}
\alias{mw_session}
\title{Default middleware.}
\usage{
mw_describe(h1)

mw_eval(h1)

mw_session(h1)
}
\description{
All default middlewares are stored within \code{middlewares} environment
inside this package. Users and add-on packages can append custom middleware
to this environment. See help page of \code{\link{middleware}} for more
details.
}
\details{
Default middlewares are:
\itemize{

\item{mw_session}{Return a handler for interactive evaluation. Provides
"eval" operation.}

\item{mw_describe}{Return a handler for session management. Supported ops
are "clone", "close" and "ls-sessions".}

\item{mw_eval}{Return a handler for interactive evaluation. Provides
"eval" operation.}

}
}

