% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{nREPL-package}
\alias{nREPL}
\alias{nREPL-package}
\title{nREPL client and server for R}
\description{
nREPL is *n*etwork REPL that provides a REPL server and client, along with
some common APIs of use to IDEs and other tools that may need to evaluate R
code in remote environments.
}
\details{
nREPL protocol was designed by Chas Emerick to be used in Clojure REPLs and
IDEs, but it's not limited or tied to to Closure in any way.
}
\author{
Vitalie Spinu
}
\references{
Full documentation and demos: \url{http://yihui.name/knitr/};
FAQ's: \url{https://github.com/yihui/knitr/blob/master/FAQ.md}
}
\seealso{
\code{\link{start_server}}, \code{\link{client}},
\code{\link{middleware}}
}

