% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrepl_middleware.R
\name{nrepl_middleware}
\alias{error_for.nREPL}
\alias{mw_describe}
\alias{mw_eval}
\alias{mw_session}
\alias{nrepl_middleware}
\alias{resp_for.nREPL}
\title{Default middleware.}
\usage{
\method{resp_for}{nREPL}(msg, ..., lst = list())

\method{error_for}{nREPL}(msg, message = "", ...,
  additional_status = list())

mw_describe(handler)

mw_session(handler)

mw_eval(handler)
}
\arguments{
\item{handler}{Handler function. See \code{\link{middleware}}.}
}
\description{
All default middlewares are stored within \code{middlewares} environment
inside this package. Users and add-on packages can append custom middleware
to this environment. See help page of \code{\link{middleware}} for more
details.
}
\details{
Default middlewares are:
\itemize{

\item{\code{mw_session}: }{Return a handler for interactive evaluation. Provides
"eval" operation.}
     
  \item{\code{mw_describe}: }{Return a handler for session management. Supported ops
are "clone", "close" and "ls-sessions".}

\item{\code{mw_eval}: }{Return a handler for interactive evaluation. Provides
"eval" operation.}

}
}

