% Generated by roxygen2 (4.0.2): do not edit by hand
\name{server}
\alias{default_handler}
\alias{server}
\alias{start_server}
\title{nREPL Server.}
\usage{
start_server(port = 4005, handler = default_handler(),
  transport_fn = transport_bencode)

default_handler(additional_middlewares = list())
}
\arguments{
\item{port}{Port number on which to start an nREPL server.}

\item{handler}{Function of variable arity to process incoming requests.}

\item{transport_fn}{Constructor that returns a transport connection
object. See \code{\link{transport}}.}

\item{additional_middlewares}{A list of middleware functions to merge into
the list of default \code{\link{middlewares}}}
}
\description{
nREPL server is a blocking connection that waits for requests from the nREPL
client and sends the responses back.
}
\details{
\code{default_handler} returns a handler which is a stack of handlers
produced by default middlewares.
}
\seealso{
\link{middlewares}
}

