% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transport}
\alias{transport}
\alias{transport_bencode}
\title{Default transports.}
\usage{
transport_bencode(con)
}
\arguments{
\item{con}{Connection object as returned by \code{\link{socketConnection}}.}
}
\description{
Transports are functions that return a list object which is a wraper arround
open \code{\link{socketConnection}}. Each transport function accept one
argument, a connection, and returns a list of at least 3 functions:
\code{write} - a function of one argument (a list request) that encodes and
then send the message to the nREPL server; \code{read} - function of one
argument (timeout) that waits for the incomming responses and decodes one
response at a time; \code{close} - function of no argment that closees the
connection.
}

