% Generated by roxygen2 (4.0.2): do not edit by hand
\name{middleware_utils}
\alias{backToTop}
\alias{backToTopError}
\alias{errorfor}
\alias{middleware_utils}
\alias{respfor}
\alias{test_middleware}
\title{Various utility functions for writing middleware.}
\usage{
respfor(msg, ..., lst = list())

errorfor(msg, message = "", ..., additional_status = list())

test_middleware(mw, ...)

backToTop()

backToTopError(message, ...)
}
\arguments{
\item{msg}{Incoming message containing \code{id} and \code{session}
elements.}

\item{...}{Key-value pairs of the response for \code{repfor} and
\code{test_middleware}. Format arguments for \code{message} in
\code{errorfor}.}

\item{lst}{A list which is merged with the elements in \code{...}}

\item{message}{Error message to be sent through.}

\item{mw}{Middleware to test}

\item{transport}{transport object as returned by \code{connect} or
\code{transport_bencode}.}

\item{msg}{Error message to send back to client on backToTop condition}
}
\description{
Various utility functions for writing middleware.
}

