% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mw_describe}
\alias{middlewares}
\alias{mw_describe}
\alias{mw_eval}
\alias{mw_session}
\title{Default middleware.}
\usage{
mw_describe(handler)

mw_session(handler)

mw_eval(handler)
}
\arguments{
\item{handler}{Handler function. See \code{\link{middleware}}.}
}
\description{
All default middlewares are stored within \code{middlewares} environment
inside this package. Users and add-on packages can append custom middleware
to this environment. See help page of \code{\link{middleware}} for more
details.
}
\details{
Default middlewares are:
\itemize{

\item{\code{mw_session}: }{Return a handler for interactive evaluation. Provides
"eval" operation.}

\item{\code{mw_describe}: }{Return a handler for session management. Supported ops
are "clone", "close" and "ls-sessions".}

\item{\code{mw_eval}: }{Return a handler for interactive evaluation. Provides
"eval" operation.}

}
}

