% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{epc_handler}
\alias{nrepl_handler}
\alias{server}
\alias{start_epc}
\alias{start_nrepl}
\title{nREPL Server.}
\usage{
start_nrepl(port = 4005, handler = nrepl_handler(),
  transport_fn = transport_bencode)

start_epc(port = 4005, handler = epc_handler(),
  transport_fn = transport_swank)

nrepl_handler(additional_middlewares = list())

epc_handler(additional_middlewares = list())
}
\arguments{
\item{port}{Port number on which to start an nREPL server.}

\item{handler}{Function of variable arity to process incoming requests.}

\item{transport_fn}{Constructor that returns a transport connection
object. See \code{\link{transport}}.}

\item{additional_middlewares}{A list of middleware functions to merge into
the list of default \code{\link{middlewares}}}
}
\description{
nREPL server is a blocking connection that waits for requests from the nREPL
client and sends the responses back.
}
\details{
\code{default_handler} returns a handler which is a stack of handlers
produced by default middlewares.
}
\seealso{
\link{middlewares}
}

